Fixed tubepy not working in Ubuntu(linux)
When I encountered the _tkinter.TclError: bitmap "assets/icons/new_tubepy_logo128.ico" not defined error, it was likely due to the fact that the iconbitmap method on Linux systems requires a bitmap file in the XBM format, not the ICO format used on Windows. Therefore, the method was unable to find and use the ICO file as the application icon.

To resolve this issue, I convert the ICO file to the XBM format and specify the file path with the @ symbol. Additionally, I needed to change the file extension to .xbm to ensure that the file was recognized as a bitmap file in the XBM format.